+++ swiftide-core/README.md
@@ -1,0 +1,3 @@
+# Swiftide-core
+
+Core crate includes low level types and traits for swiftide that are used by other crates.
