+++ swiftide-indexing/src/lib.rs
