+++ swiftide-integrations/src/aws_bedrock/models/anthropic.rs
