+++ swiftide-integrations/src/aws_bedrock/models/mod.rs
