+++ swiftide-integrations/src/openai/embed.rs
@@ -2,7 +2,7 @@ use anyhow::{Context as _, Result};
 use async_openai::types::CreateEmbeddingRequestArgs;
 use async_trait::async_trait;
 
-use crate::{EmbeddingModel, Embeddings};
+use swiftide_core::{EmbeddingModel, Embeddings};
 
 use super::OpenAI;
