+++ swiftide-integrations/src/openai/mod.rs
