+++ swiftide-integrations/src/openai/simple_prompt.rs
@@ -1,7 +1,7 @@
 //! This module provides an implementation of the `SimplePrompt` trait for the `OpenAI` struct.
 //! It defines an asynchronous function to interact with the `OpenAI` API, allowing prompt processing
 //! and generating responses as part of the Swiftide system.
-use crate::{prompt::Prompt, SimplePrompt};
+use swiftide_core::{prompt::Prompt, SimplePrompt};
 use async_openai::types::{ChatCompletionRequestUserMessageArgs, CreateChatCompletionRequestArgs};
 use async_trait::async_trait;
