+++ swiftide-integrations/src/scraping/loader.rs
@@ -2,10 +2,7 @@ use derive_builder::Builder;
 use spider::website::Website;
 use tokio::{runtime::Handle, sync::RwLock};
 
-use crate::{
-    indexing::{IndexingStream, Node},
-    Loader,
-};
+use swiftide_core::{indexing_stream::IndexingStream, node::Node, traits::Loader};
 
 #[derive(Debug, Builder)]
 #[builder(pattern = "owned")]
