+++ swiftide/Cargo.toml
@@ -26,82 +26,42 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 strum = { workspace = true }
 strum_macros = { workspace = true }
+indoc = { workspace = true }
 
-chrono = { version = "0.4" }
 ignore = "0.4"
-indoc = "2.0"
 text-splitter = { version = "0.14", features = ["markdown"] }
 
 # Local dependencies
 swiftide-core = { path = "../swiftide-core" }
-
-# Integrations
-async-openai = { version = "0.23", optional = true }
-qdrant-client = { version = "1.10", optional = true, default-features = false, features = [
-  "serde",
-] }
-redis = { version = "0.25", features = [
-  "aio",
-  "tokio-comp",
-  "connection-manager",
-  "tokio-rustls-comp",
-], optional = true }
-tree-sitter = { version = "0.22", optional = true }
-tree-sitter-rust = { version = "0.21", optional = true }
-tree-sitter-python = { version = "0.21", optional = true }
-tree-sitter-ruby = { version = "0.21", optional = true }
-tree-sitter-typescript = { version = "0.21", optional = true }
-tree-sitter-javascript = { version = "0.21", optional = true }
-fastembed = { version = "3.6", optional = true }
-spider = { version = "1.98", optional = true }
-htmd = { version = "0.1", optional = true }
-aws-config = { version = "1.5", features = [
-  "behavior-version-latest",
-], optional = true }
-aws-credential-types = { version = "1.2", features = [
-  "hardcoded-credentials",
-], optional = true }
-aws-sdk-bedrockruntime = { version = "1.37", features = [
-  "behavior-version-latest",
-], optional = true }
-secrecy = { version = "0.8.0", optional = true }
-reqwest = { version = "0.12.5", optional = true, default-features = false }
+swiftide-integrations = { path = "../swiftide-integrations" }
 
 [features]
-default = ["rustls"]
+default = []
 all = ["qdrant", "redis", "tree-sitter", "openai", "fastembed"]
-# Ensures rustls is used 
-rustls = ["reqwest/rustls-tls-native-roots"]
 # Qdrant for storage
-qdrant = ["dep:qdrant-client"]
+qdrant = ["swiftide-integrations/qdrant"]
 # Redis for caching and storage
-redis = ["dep:redis"]
+redis = ["swiftide-integrations/redis"]
 # Tree-sitter for code operations and chunking
-tree-sitter = [
-  "dep:tree-sitter",
-  "dep:tree-sitter-rust",
-  "dep:tree-sitter-python",
-  "dep:tree-sitter-ruby",
-  "dep:tree-sitter-typescript",
-  "dep:tree-sitter-javascript",
-]
+tree-sitter = ["swiftide-integrations/tree-sitter"]
 # OpenAI for embedding and prompting
-openai = ["dep:async-openai"]
+openai = ["swiftide-integrations/openai"]
 # Groq prompting
-groq = ["dep:async-openai", "dep:secrecy", "dep:reqwest"]
+groq = ["swiftide-integrations/groq"]
 # FastEmbed (by qdrant) for fast, local embeddings
-fastembed = ["dep:fastembed"]
+fastembed = ["swiftide-integrations/fastembed"]
 # Scraping via spider as loader and a html to markdown transformer
-scraping = ["dep:spider", "dep:htmd"]
+scraping = ["swiftide-integrations/scraping"]
 # AWS Bedrock for prompting
-aws-bedrock = [
-  "dep:aws-config",
-  "dep:aws-credential-types",
-  "dep:aws-sdk-bedrockruntime",
-]
+aws-bedrock = ["swiftide-integrations/aws-bedrock"]
 
 [dev-dependencies]
 swiftide-core = { path = "../swiftide-core", features = ["test-utils"] }
+
+async-openai = { workspace = true }
+qdrant-client = { workspace = true, default-features = false, features = [
+  "serde",
+] }
 
 test-log = { workspace = true }
 testcontainers = { workspace = true }
@@ -111,18 +71,8 @@ wiremock = { workspace = true }
 test-case = { workspace = true }
 insta = { workspace = true }
 
-[lints.rust]
-unsafe_code = "forbid"
-
-[lints.clippy]
-cargo = { level = "warn", priority = -1 }
-pedantic = { level = "warn", priority = -1 }
-blocks_in_conditions = "allow"
-must_use_candidate = "allow"
-module_name_repetitions = "allow"
-missing_fields_in_debug = "allow"
-# Should be fixed asap
-multiple_crate_versions = "allow"
+[lints]
+workspace = true
 
 [package.metadata.docs.rs]
 all-features = true
